+++ scripts/benchmarking/run_benchmarks_genn.sh
@@ -1,12 +1,18 @@
 #!/bin/bash -e
 
+if [ "$1" = "COBAHH_MON.py" ]; then 
+    mon=0.1;
+else
+    mon="";
+fi
+    
 for spikemon in "true"; do
    for threads in 0 -1; do
        for scaling in 0.05 0.1 0.25 0.5 1 2 4 8 16 32; do
 	   time=1;
                for repeat in 1 2 3 4 5; do
                    echo Repeat $repeat
-                   python $1 $scaling genn $threads $spikemon true $time
+                   python $1 $scaling genn $threads $spikemon true $time $mon
                    rm -r GeNNworkspace
 	       done
        done
@@ -14,13 +20,15 @@ for spikemon in "true"; do
 done
 		   
 # The really long runs (don't run with GeNN CPU-only, etc.)
-for spikemon in "true"; do
-    for scaling in 64 128 256 512; do
-	time=1;
+if [ "$1" = "Mbody_example.py" ]; then 
+    for spikemon in "true"; do
+	for scaling in 64 128 256 512; do
+	    time=1;
 	    for repeat in 1 2 3; do
 		echo Repeat $repeat
-                python $1 $scaling genn 0 $spikemon true $time
+                python $1 $scaling genn 0 $spikemon true $time $mon
                 rm -r GeNNworkspace
 	    done
+	done
     done
-done
+fi
